+++ lib/sidekiq/worker/job/extension.rb
@@ -1,0 +1,10 @@
+module Sidekiq
+  module Worker
+    class Job
+      module Extension
+        # @attr_accessor job [Sidekiq::Worker::Job]
+        attr_accessor :job
+      end
+    end
+  end
+end